Support .env files for environment variables
This change adds support for loading of environment variables from .env
files, instead of from an environment.json file.

Currently as part of the artifact release command, a required "vars"
argument must point to a JSON file containing environment variables.
This file is stored in the release artifact under the filename
environment.json, regardless of its original filename.

As part of DFD runtime, an "environment" path key defaults to point to
this file ("environment.json"). This path can be overloaded with an
alternate JSON file if desired, by passing it into the release command's
paths input.

At runtime, as part of the environment modifications that DFD does to
all Python code that run under it, this "environment" path key is used
to locate the input JSON file, and this file is loaded into the
environment.

This commit makes two changes: first, the release command no longer
_requires_ that the user pass an environment JSON file. Second, at
runtime, if the "environment" path does not end in ".json", the input
file is assumed to be in .env format (i.e. key=value), and the
django-dotenv package is used to read in the file.

This allows users to use .env files instead of environment.json files,
while maintaining all existing functionality.